Codelab: minikafka binary reader support read length prefixed strings

+++ one/codelab/minikafka/api_versions_request_data.cc
@@ -88,7 +88,8 @@ absl::Status ApiVersionsRequestData::ParseFrom(KafkaBinaryReader* reader,
       }
 
       std::string data;
-      ONE_RETURN_IF_NOT_OK(reader->ReadString(&data, length));
+      ONE_RETURN_IF_NOT_OK(
+          reader->ReadString(&data, static_cast<int32_t>(length)));
     }
   }
